#8 feat(paginationWrapper): рефакторинг + новый запрос для последних новостей для баннеров
