+++ src/components/BannersList/styles.module.css
@@ -4,4 +4,6 @@
   width: 100%;
   gap: 12px;
   box-sizing: border-box;
+  max-height: 1200px;
+  overflow-y: auto;
 }
