+++ src/components/PaginationWrapper/PaginationWrapper.jsx
@@ -1,0 +1,15 @@
+import Pagination from "@/components/Pagination";
+
+const PaginationWrapper = (props) => {
+  const { top, bottom, children, ...paginationProps } = props;
+
+  return (
+    <>
+      {top && <Pagination {...paginationProps} />}
+      {children}
+      {bottom && <Pagination {...paginationProps} />}
+    </>
+  );
+};
+
+export default PaginationWrapper;
