+++ src/components/PaginationWrapper/index.js
@@ -1,0 +1,3 @@
+import PaginationWrapper from "./PaginationWrapper";
+
+export default PaginationWrapper;
