+++ src/pages/Main/Main.jsx
@@ -1,41 +1,13 @@
 import styles from "./styles.module.css";
 
-import { getNews } from "@/api/apiNews";
-
-import { useDebounce } from "@/helpers/hooks/useDebounce";
-import { useFetch } from "@/helpers/hooks/useFetch";
-import { useFilters } from "@/helpers/hooks/useFilers";
-
-import { PAGE_SIZE } from "@/constants/constants";
-
 import LatestNews from "@/components/LatestNews";
 import NewsByFilters from "@/components/NewsByFilters";
 
 const Main = () => {
-  const { filters, changeFilter } = useFilters({
-    page_number: 1,
-    page_size: PAGE_SIZE,
-    category: null,
-    keywords: "",
-  });
-
-  const debouncedKeywords = useDebounce(filters.keywords, 1500);
-
-  const { data, isLoading } = useFetch(getNews, {
-    ...filters,
-    keywords: debouncedKeywords,
-  });
-
   return (
     <main className={styles.root}>
-      <LatestNews isLoading={isLoading} banners={data && data.news} />
-
-      <NewsByFilters
-        news={data?.news}
-        isLoading={isLoading}
-        filters={filters}
-        changeFilter={changeFilter}
-      />
+      <LatestNews />
+      <NewsByFilters />
     </main>
   );
 };
